feat(context): add UserContext for session management and integrate into layout and post components
