+++ packages/frontend_app/src/context/UserContext/index.tsx
@@ -1,0 +1,41 @@
+"use client";
+
+import type { Session } from "@supabase/supabase-js";
+import { createContext, useContext, useEffect, useState } from "react";
+import { createClient } from "#src/supabase/client";
+
+type Context = {
+  session: Session | null | undefined;
+};
+
+const UserContext = createContext<Context | undefined>(undefined);
+
+type Props = {
+  children: React.ReactNode;
+};
+
+export const UserContextProvider = ({ children }: Props) => {
+  const [session, setSession] = useState<Session | null | undefined>(undefined);
+
+  useEffect(() => {
+    const supabase = createClient();
+    const {
+      data: { subscription },
+    } = supabase.auth.onAuthStateChange((_event, nextSession) => {
+      setSession(nextSession);
+    });
+    return () => subscription.unsubscribe();
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ session }}>{children}</UserContext.Provider>
+  );
+};
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (!context) {
+    throw new Error("useUserContext must be used within a UserContextProvider");
+  }
+  return context;
+};
